third_party/WebKit/LayoutTests/accessibility/ignore-text-elements-for-menu-item-related-roles-expected.txt

menu item checkbox
menu item radio
PASS !menu.childAtIndex(0).childAtIndex(0) is true
PASS !menu.childAtIndex(1).childAtIndex(0) is true
AXTitle: menu item checkbox
AXTitle: menu item radio
PASS successfullyParsed is true

TEST COMPLETE
